chore(suite): move fiat rates to common
