+++ packages/env-utils/src/envUtils.native.ts
@@ -9,6 +9,8 @@ const isWeb = () => false;
 
 const isDesktop = () => false;
 
+const isNative = () => true;
+
 const getEnvironment = () => 'mobile' as const;
 
 const getUserAgent = () => '';
@@ -74,6 +76,7 @@ const getOsFamily = (): 'Linux' => 'Linux';
 export const envUtils: EnvUtils = {
     isWeb,
     isDesktop,
+    isNative,
     getEnvironment,
     getUserAgent,
     isAndroid,
