+++ packages/env-utils/src/envUtils.ts
@@ -6,6 +6,8 @@ export const isWeb = () => process.env.SUITE_TYPE === 'web';
 
 export const isDesktop = () => process.env.SUITE_TYPE === 'desktop';
 
+export const isNative = () => false;
+
 export const getEnvironment = (): Environment => {
     if (isWeb()) return 'web';
     return 'desktop';
@@ -119,6 +121,7 @@ const getDeviceType = () => getUserAgentParser().getDevice().type;
 export const envUtils: EnvUtils = {
     isWeb,
     isDesktop,
+    isNative,
     getEnvironment,
     getUserAgent,
     isAndroid,
