+++ suite-common/wallet-core/src/fiat-rates/fiatRatesMiddleware.ts
@@ -1,76 +1,39 @@
-import { createMiddlewareWithExtraDeps } from '@suite-common/redux-utils';
-import { BLOCKCHAIN as TREZOR_CONNECT_BLOCKCHAIN_ACTIONS } from '@trezor/connect';
-import { isTickerSupportedByBlockbook } from '@suite-common/fiat-services';
+import { isAnyOf } from '@reduxjs/toolkit';
 
-import { transactionsActions } from '../transactions/transactionsActions';
-import { accountsActions } from '../accounts/accountsActions';
-import { selectAccounts } from '../accounts/accountsReducer';
+import { isNative } from '@trezor/env-utils';
+import { createMiddlewareWithExtraDeps } from '@suite-common/redux-utils';
+
 import {
-    removeFiatRatesForDisabledNetworksThunk,
-    updateCurrentFiatRatesThunk,
-    initFiatRatesThunk,
-    updateLastWeekFiatRatesThunk,
+    fetchFiatRatesThunk,
+    updateMissingTxFiatRatesThunk,
     updateTxsFiatRatesThunk,
-    onUpdateFiatRateThunk,
 } from './fiatRatesThunks';
 import { blockchainActions } from '../blockchain/blockchainActions';
-import { selectAccountTransactions } from '../transactions/transactionsReducer';
+import { accountsActions } from '../accounts/accountsActions';
+import { transactionsActions } from '../transactions/transactionsActions';
 
 export const prepareFiatRatesMiddleware = createMiddlewareWithExtraDeps(
     (action, { dispatch, extra, next, getState }) => {
         const {
+            actions: { setWalletSettingsLocalCurrency },
             selectors: { selectLocalCurrency },
-            actions: { setWalletSettingsLocalCurrency, changeWalletSettingsNetworks },
         } = extra;
-        const prevStateAccounts = selectAccounts(getState());
-        const prevStateLocalCurrency = selectLocalCurrency(getState());
 
-        if (accountsActions.updateAccount.match(action)) {
-            // fetch rates for new tokens added on account update
-            const account = action.payload;
-            const prevAccount = prevStateAccounts.find(
-                a => a.descriptor === account.descriptor && a.symbol === account.symbol,
+        if (isAnyOf(accountsActions.updateAccount, accountsActions.createAccount)(action)) {
+            dispatch(
+                fetchFiatRatesThunk({
+                    rateType: 'current',
+                    localCurrency: selectLocalCurrency(getState()),
+                }),
             );
-
-            if (account.tokens) {
-                const difference = account.tokens.filter(
-                    token =>
-                        !prevAccount?.tokens?.find(
-                            prevToken => prevToken.contract === token.contract,
-                        ),
-                );
-
-                difference.forEach(token => {
-                    if (token.symbol) {
-                        dispatch(
-                            updateCurrentFiatRatesThunk({
-                                ticker: {
-                                    symbol: account.symbol,
-                                    tokenAddress: token.contract,
-                                },
-                            }),
-                        );
-                    }
-                });
-            }
-        }
-
-        if (accountsActions.createAccount.match(action)) {
-            // fetch current rates for account's tokens
-            const account = action.payload;
-            account.tokens?.forEach(token => {
-                if (!token.symbol) {
-                    return;
-                }
+            if (!isNative()) {
                 dispatch(
-                    updateCurrentFiatRatesThunk({
-                        ticker: {
-                            symbol: account.symbol,
-                            tokenAddress: token.contract,
-                        },
+                    fetchFiatRatesThunk({
+                        rateType: 'lastWeek',
+                        localCurrency: selectLocalCurrency(getState()),
                     }),
                 );
-            });
+            }
         }
 
         if (transactionsActions.addTransaction.match(action)) {
@@ -80,42 +43,36 @@ export const prepareFiatRatesMiddleware = createMiddlewareWithExtraDeps(
                 updateTxsFiatRatesThunk({
                     account,
                     txs: transactions,
-                    localCurrency: prevStateLocalCurrency,
+                    localCurrency: selectLocalCurrency(getState()),
                 }),
             );
         }
 
         if (setWalletSettingsLocalCurrency.match(action)) {
-            // For coins depending on Coingecko API we always download fiat rates for only one currency.
             const { localCurrency } = action.payload;
-
-            dispatch(updateLastWeekFiatRatesThunk(localCurrency));
-
-            prevStateAccounts.forEach(account => {
-                // if fiat rates are downloaded from blockbook we already have all the currency-rate pairs.
-                if (isTickerSupportedByBlockbook({ symbol: account.symbol })) return;
-
-                const accountTransactions = selectAccountTransactions(getState(), account.key);
+            // We need to pass localCurrency as a parameter, because it is not yet updated in the store
+            dispatch(fetchFiatRatesThunk({ rateType: 'current', localCurrency }));
+            if (!isNative()) {
                 dispatch(
-                    updateTxsFiatRatesThunk({
-                        account,
-                        txs: accountTransactions,
-                        localCurrency,
+                    fetchFiatRatesThunk({
+                        rateType: 'lastWeek',
+                        localCurrency: selectLocalCurrency(getState()),
                     }),
                 );
-            });
-        }
-
-        if (changeWalletSettingsNetworks.match(action)) {
-            dispatch(removeFiatRatesForDisabledNetworksThunk());
+            }
+            dispatch(updateMissingTxFiatRatesThunk({ localCurrency }));
         }
 
         if (blockchainActions.connected.match(action)) {
-            dispatch(initFiatRatesThunk(action.payload));
-        }
-
-        if (action.type === TREZOR_CONNECT_BLOCKCHAIN_ACTIONS.FIAT_RATES_UPDATE) {
-            dispatch(onUpdateFiatRateThunk(action.payload));
+            // TODO: verify if this is necessary, it should work only based on addTransaction action
+            // just to be safe, refetch historical rates for transactions stored without these rates
+            // dispatch(updateMissingTxFiatRatesThunk());
+            dispatch(
+                fetchFiatRatesThunk({
+                    rateType: 'current',
+                    localCurrency: selectLocalCurrency(getState()),
+                }),
+            );
         }
 
         return next(action);
