+++ suite-common/wallet-core/src/index.ts
@@ -8,8 +8,9 @@ export * from './blockchain/blockchainActions';
 export * from './blockchain/blockchainReducer';
 export * from './blockchain/blockchainThunks';
 export * from './blockchain/blockchainMiddleware';
-export * from './fiat-rates/fiatRatesActions';
 export * from './fiat-rates/fiatRatesReducer';
+export * from './fiat-rates/fiatRatesActions';
+export * from './fiat-rates/fiatRatesSelectors';
 export * from './fiat-rates/fiatRatesThunks';
 export * from './fiat-rates/fiatRatesMiddleware';
 export * from './discovery/discoveryActions';
