+++ suite-native/accounts/tsconfig.json
@@ -20,7 +20,6 @@
         },
         { "path": "../atoms" },
         { "path": "../ethereum-tokens" },
-        { "path": "../fiat-rates" },
         { "path": "../formatters" },
         { "path": "../forms" },
         { "path": "../module-settings" },
