+++ suite-native/app/tsconfig.json
@@ -29,7 +29,6 @@
         { "path": "../atoms" },
         { "path": "../biometrics" },
         { "path": "../config" },
-        { "path": "../fiat-rates" },
         { "path": "../message-system" },
         { "path": "../module-accounts" },
         { "path": "../module-accounts-import" },
