+++ suite-native/assets/tsconfig.json
@@ -21,7 +21,6 @@
         { "path": "../accounts" },
         { "path": "../atoms" },
         { "path": "../ethereum-tokens" },
-        { "path": "../fiat-rates" },
         { "path": "../formatters" },
         { "path": "../module-settings" },
         { "path": "../navigation" },
