--- suite-native/fiat-rates/package.json
@@ -1,27 +1,0 @@
-{
-    "name": "@suite-native/fiat-rates",
-    "version": "1.0.0",
-    "private": true,
-    "license": "See LICENSE.md in repo root",
-    "sideEffects": false,
-    "main": "src/index",
-    "scripts": {
-        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
-        "test:unit": "jest -c ../../jest.config.base.js --passWithNoTests",
-        "type-check": "tsc --build"
-    },
-    "dependencies": {
-        "@mobily/ts-belt": "^3.13.1",
-        "@reduxjs/toolkit": "1.9.5",
-        "@suite-common/fiat-services": "workspace:*",
-        "@suite-common/redux-utils": "workspace:*",
-        "@suite-common/suite-config": "workspace:*",
-        "@suite-common/wallet-config": "workspace:*",
-        "@suite-common/wallet-core": "workspace:*",
-        "@suite-common/wallet-types": "workspace:*",
-        "@suite-common/wallet-utils": "workspace:*",
-        "@trezor/connect": "workspace:*",
-        "date-fns": "^2.30.0",
-        "proxy-memoize": "2.0.2"
-    }
-}
