--- suite-native/fiat-rates/src/index.ts
@@ -1,7 +1,0 @@
-export * from './fiatRatesReducer';
-export * from './fiatRatesActions';
-export * from './fiatRatesSelectors';
-export * from './fiatRatesThunks';
-export * from './fiatRatesMiddleware';
-export * from './types';
-export * from './utils';
