--- suite-native/fiat-rates/tsconfig.json
@@ -1,28 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": { "outDir": "libDev" },
-    "references": [
-        {
-            "path": "../../suite-common/fiat-services"
-        },
-        {
-            "path": "../../suite-common/redux-utils"
-        },
-        {
-            "path": "../../suite-common/suite-config"
-        },
-        {
-            "path": "../../suite-common/wallet-config"
-        },
-        {
-            "path": "../../suite-common/wallet-core"
-        },
-        {
-            "path": "../../suite-common/wallet-types"
-        },
-        {
-            "path": "../../suite-common/wallet-utils"
-        },
-        { "path": "../../packages/connect" }
-    ]
-}
