+++ suite-native/module-accounts/tsconfig.json
@@ -24,7 +24,6 @@
         { "path": "../analytics" },
         { "path": "../atoms" },
         { "path": "../ethereum-tokens" },
-        { "path": "../fiat-rates" },
         { "path": "../formatters" },
         { "path": "../forms" },
         { "path": "../graph" },
