+++ suite-native/state/package.json
@@ -20,7 +20,6 @@
         "@suite-common/wallet-core": "workspace:*",
         "@suite-native/config": "workspace:*",
         "@suite-native/device": "workspace:*",
-        "@suite-native/fiat-rates": "workspace:*",
         "@suite-native/graph": "workspace:*",
         "@suite-native/logger": "workspace:*",
         "@suite-native/message-system": "workspace:*",
